+++ src/components/Widgets/Banner/IdeaBannerWidget.svelte
@@ -68,7 +68,6 @@
         };
     }
     $: $nostrCache, fetchIdea();
-
     $: $nostrManager, initialize();
 </script>
