+++ src/helperStore.js
@@ -1,5 +1,10 @@
 // helperStore.js
-import { writable } from 'svelte/store';
+import { writable, derived } from 'svelte/store';
 
 export const helperStore = writable(null);
 export const sidebarOpen = writable(false);
+
+
+export const contentContainerClass = derived(sidebarOpen, $sidebarOpen =>
+    $sidebarOpen ? "combined-content-container sidebar-open" : "combined-content-container"
+);
