+++ src/views/JobManager.svelte
@@ -1,24 +1,14 @@
 <!-- JobManagerView.svelte -->
 <script>
   import { onMount } from "svelte";
-  import Menu from "../components/Menu.svelte";
+  import Menu from "../components/Sidebar/Sidebar.svelte";
   import JobManager from "../components/JobManager2/JobManager.svelte";
-  import Footer from "../components/Footers/FooterBS.svelte";
-  import { sidebarOpen } from "../helperStore.js";
+  import Footer from "../components/Footers/Footer.svelte";
+  import { contentContainerClass } from "../helperStore.js";
   import Banner from "../components/Banner.svelte";
-  import ToolBar from "../components/ToolBar.svelte";
+  import ToolBar from "../components/Toolbar/Toolbar.svelte";
 
   onMount(async () => {});
-
-  let contentContainerClass = "combined-content-container";
-
-  $: {
-    if ($sidebarOpen) {
-      contentContainerClass = "combined-content-container sidebar-open";
-    } else {
-      contentContainerClass = "combined-content-container";
-    }
-  }
 
   let bannerImage = "../../img/Banner1u.png";
   let title = "BitSpark";
@@ -30,7 +20,7 @@
   <div class="flex-grow">
     <Banner {bannerImage} {title} {subtitle} show_right_text={false} />
     <ToolBar />
-    <div class={contentContainerClass}>
+    <div class={$contentContainerClass}>
       <JobManager />
     </div>
   </div>
@@ -49,10 +39,4 @@
   .flex-grow {
     flex-grow: 1;
   }
-
-  .combined-content-container {
-    padding: 1rem;
-    margin: 1rem auto;
-  }
 </style>
-
